fix: tags on load
