+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -11,6 +11,8 @@
 * [ ] patch
 
 # Further Information (screenshots, bug report links, etc.)
+<!-- If this PR fixes an open issue, please write 'Closes #0' or 'Fixes #0',
+where 0 is the issue number -->
 
 # Checklist
